* Refactoring * Test fixing

+++ src/main/java/de/tubs/cs/ias/asm_test/TaintStringHelper.java
@@ -39,7 +39,7 @@ public class TaintStringHelper {
         List<IASTaintRange> ranges = new ArrayList<>(count);
 
         for (int start = 0, end = width, i = 0; i < count; start += width, end += width, i++) {
-            ranges.add(new IASTaintRange(start, end, (short) (i + 1)));
+            ranges.add(new IASTaintRange(start, end, (i + 1)));
         }
 
         return ranges;
